evdev-remapd: Implement clean shutdown of event loop
asyncio is confusing. It took a while to work out how to correctly
use the shutdown() method to cleanly handle a KeyboardInterrupt as
well as a SIGTERM.

Note that the shutdown() method is annotated as a coroutine because
making it a native async method leads to a warning on exit if you
kill it with a KeyboardInterrupt. In this case the signal handler
is never invoked so the async function is never awaited. It's
harmless but ugly. Anotated coroutines don't have this behaviour.