evdev-remapd: Handle remappings in a more generic way
This is the first step towards proper generic remappings. To that
end there are two main parts to this change:

1) Only a single uinput device is created - it mirrors the selected
physical device's capabilities plus whatever else is needed to
achieve the remapping (so adding keys that didn't otherwise exist)

2) All remappings are treated the same - so the event code is
changed (and multiple events are injected for 1:N mappings) but
everything else stays the same. Particularly, this means that a
held-down key on the input side always maps to a held down key
on the output side.