evdevremapkeys: Differentiate identically named devices
Sometimes we see multiple devices with the same name. This often
happens with keyboards, where two evdev devices are present, with
the various keys split between them.

The old code would pick the first device that matched the name,
but it might not be the one you want.

Now you can differentiate them by specifying the physical location
of the device or the device filename.

I've expanded the example configuration to illustrate this, and
also to show how to specify remappings for multiple devices.