evdevremapkeys: Add --read-events option
Add a helper mode to print out events from a device. This can help
you to identify which events are generated by which physical inputs.

It has value beyond just running evtest as it doesn't print out events
that can't be remapped (like axis events) which would otherwise generate
a lot of noise.