evdevremapkeys: Handle hotplug events for devices
Today, we can only handle devices if they are present when we start
the program, and if they are unplugged, we lose them until the
program is restarted.

Rather, let's listen for udev events and try to pick up devices
when they are plugged in. The current logic is a bit crude, in
that it will scan for all known devices whenever an input device
is added, but it does the job.