Rename meta variable for --read-events
When using --read-events you are selecting a device, not an event.
At first I only read "--read-events EVENT_ID" which lead me to believe that I
needed to supply a (key) event when in fact it listens for all events on the
selected device.
The help reflects this already, but I did not read it at first. I think DEVICE
fits better to the actual parameter.