Add heading to --list-devices
`--list-devices` is used to find out how to filter for an input. The config file
takes input_fn, input_phys and input_name. You can list these by using
`--list-devices` but it does not state which is which.

This adds a heading to reflect the values.
The number of spaces after "input_fn" is not random. Since I expect the values
of input_fn to be at somthing like `/dev/input/event0` or longer the this length
was chosen:
* input_fn:         \t
* /dev/input/event0:\t