build: switch to uv to manage our build
This has been a long time coming, but I have to sort out the build and
packaging story, and now that `uv` is a thing, it provides an easy way
to do this.

This change heavily streamlines how building and packaging happens.

* Switch to pyproject.toml and retire setup.cfg
* Use uv to manage dependencies and to run tools
* Switch to hatch for building the sdist, and retire setuptools
* Switch to ruff for linting and formatting

The minimum python version is also updated to 3.8 as the more modern
versions of pyinstaller require it. It seems like a reasonable thing to do
given how much time has passed, in any case.